Lexical analyzer main file
The program itself takes a file and reads its contents. While it does that, it tries to recognize different sequences of tokens which is useful if you want a program that requires these tokens to be defined for the program to change functionality. It can definitely use some more work though.